Rewrite the DoublyLinkedList into an unsafe version, add a NodePool, and replace all locks by a global MM_LOCK.
The DoublyLinkedList in Safe Rust turned out to be buggy and harder for debugging.
Therefore, it has been replaced by an unsafe version with proven algorithms in this branch.
The goal is still to revive the safe version later.

Memory Management has been hardened further by introducing a single global MM_LOCK instead of finer ones per MM component.
This simplifies the code and reliably protects against deadlocks.

Finally, allocating or deallocating single FreeList entries inside its allocate and deallocate functions is not stable.
When running a concurrent memory manager torture test on each application processor inside the kernel, this can lead to a recursive mess corrupting the FreeList.
Therefore, a new approach has been taken based on the following:
* We can estimate that an allocate operation leaves the node count inside the FreeList unchanged or removes a single node, but not more.
* Respectively, a deallocate operation may leave the node count unchanged, add a single node or remove one.
* MM uses two FreeLists (physical and virtual memory) at the same time.

Combining this information, we can design a NodePool that serves as a dump for removed nodes from the FreeList.
For deallocate operations, we have to ensure that the pool always contains at least 2 nodes. They can be allocated when it is safe to do so.
As a side effect, this NodePool also enables deallocate to reuse a node that was previously removed. This saves memory allocations.

The new code also implements reuniting blocks in the FreeList.
After testing it with a concurrent memory manager torture test on 64 CPUs in QEMU, no crashes occur anymore and the FreeLists remain in a good state.